refactor: 리팩로링중

+++ prisma/migrations/20251019105454_add_profile_id_to_media/migration.sql
@@ -5,4 +5,4 @@
 
 */
 -- AlterTable
-ALTER TABLE "Media" ADD COLUMN     "profile_id" BIGINT NOT NULL;
+ALTER TABLE "Media" ADD COLUMN     "profile_id" INT NOT NULL;
